Issue #69: Fixed SenderError in register example
A BeskarError is being raised as a result of SenderError in the
async_sender library when running the register example; this is because
the Message being sent does not have the to, cc or bcc attributes as
consequence of being reinitialised.

This was resolved by directly adding the Message arguements to the
`self.app.ctx.mail.send_message()` method.

Also:
* Changed the `reply_to` value from a list to a str
* Changed the response on the `/register` endpoint to JSONResponse